Got rid of BufferedWriter protocol
I'm not currently convinced that the distinction between a Writer and a
BufferedWriter is something that should be defined at the protocol level, though
this is mostly based off of the fact that `clojure.java.io` doesn't define a
difference in the protocols, instead relying on implementations to capture the
difference.

`clojure.java.io` also doesn't define how to interact with the various readers
and streams, it only defines a way to create them.

From this perspective, I'm going to default to buffered implementations for the
time being, and move over to having a clearer distinction if the need arises.

It's also important to note that it may be hard/impossible to reliably create
and work with a buffered or unbuffered reader/writer/stream, depending on the
platform. For example, Node doesn't have an unbuffered reader or writer (that I
could find), instead relying on an internal buffer (and the ability to define
your own buffer size), as well as return values and signals to define when the
buffer is full and undergoing a flush or in need of a `.read`.

This same distinction comes about when trying to read/write/stream in a
synchronous fashion; node does not make this simple, and in some cases -- such
as reading from an input stream -- there's no reliable way to synchronously
read, as reading from an uninitialized buffer returns `null` and so does reading
from a closed readable stream. So instead of blocking until there's data, it
just returns immediately. I'll write more on this in a later commit.